NFL-Network archive - Season 2014 added
frostyon420 reported problems with the show 'Top 100 Players'. The
problem was, that the current season could not be played.

Now I determine the current season and removed the fixed '2013'
references and check it on the current season variable.
So next season only the show_archives need to be update with the new
year and new 'CIDs'.