Fix manifest retrieval for URLs without explicit port
For some reason, some manifest URLs are returned with "443" in them, and
others without any port. The original code assumed there would /always/
be port 443 specified.

This is particularly odd in 2013 week 3. Lions vs Redskins return a
manifest without a port, but Buccaneers vs Patriots /does/ return a
manifest URL with a port.