Small performance gain for check_for_subscription()
Since we aren't /actually/ returning the subscription type, it makes no
sense to parse the XML and then search the dict.

On the rpi, parsing and searching takes 0.1s. On a fresh login, we call
this twice. Since this new check is basically instantaneous, rpi users
get a 0.2s speedup.

In the future, if/when we support alternate subscription types
(team only, etc), then we may need to go back to parsing the XML.
Until then, let's take this performance win.