Rip out caching
Every time the skin asks for season-and-week data, it'll get it fresh.
We'll make it the skin's job to cache. Thankfully, this is a non issue
since the skin keeps that data in memory while running.

The seasons and week data doesn't change frequently, so there's no
problem with not checking if its expired. Restarting the addon will
requests new data, so we now no longer have caching bugs/problems.

This also removes plugin.cache (StorageServer) as a dependency.

As a side-effect of this work, the Hall of Fame game is appropriately
labeled.