bug fix: restore headers to m3u8 stream manifests
I am not certain this is needed; which headers we do and don't need needs to be
audited.

We also need a better way having the user override the headers. Allowing a sane
default (since sometimes we need to use certain settings to access the content),
but also exposing more control.