get_weeks_games() -> get_games(); refactor and add tests
Another make_request() down.

I was going to mess with the return values here, as returning the raw server
response is just not ok, but I'm beginning to develop an idea for a better API,
and the effort expended here would be mostly just thrown away. So we're just
going to roll with this substandard setup for now.