retire get_streams(); create get_[game/redzone/nfl_network]_streams()
There's some code duplication here, to be sure, but it's a hell of a lot easier
to follow when tracing one of them.

I /really/ hope there's an easier way to snake through the akamai API and get
the results we want. But now, we have tests, so it should be a lot easier to
attempt.