kill make_request()/parse_response(); move show acquisition to data.py
The show acquisition code is broken at the moment, and will be entirely
rewritten, anyway, which is why the make_requests stuff is being ripped out now
rather than waiting.

I'm impatient.

m3u8_to_dict was updated to no longer use make_request(). It will likely move,
but that will come later with the introduction of the stream class.